Blokada baz podłączonych do Edytora opisu taksacyjnego
Dopóki Edytor opisu trzyma połączenie z bazą, żaden inny skrypt wtyczki
nie może się z nią połączyć (także do odczytu). Kontrola w Baza.polacz,
Baza.usun_kwerendy (DAO) i aktualizacja_upul.core.db.connect; Anonimizuj
sprawdza blokadę przed kopią X.mdb -> X_BDO.mdb. Komunikat krytyczny i
przerwanie skryptu bez zmian w bazie. Niesprawdzane wywołania polacz()
uzupełnione o wyjście przy odmowie połączenia.

+++ skrypty/baza_rozlicz_pow_wydz.py
@@ -578,7 +578,8 @@ class RozliczPowierzchnieWydz(SprawdzWydzielenia):
         plik.close()
 
     def wyczysc_arod_land_use(self):
-        self.baza.polacz()
+        if not self.baza.polacz():
+            return False
         self.baza.wpisz("DELETE FROM F_AROD_LAND_USE")
         QgsMessageLog.logMessage(
             'Wyczyszczono tabelę F_AROD_LAND_USE', 'Las-R', Qgis.Info)
@@ -600,7 +601,8 @@ class RozliczPowierzchnieWydz(SprawdzWydzielenia):
 
         self.baza.zamknij()
         self.baza.utworz_kopie('kopia_rozliczPow')
-        self.baza.polacz()
+        if not self.baza.polacz():
+            return False
 
         # sl tworzony w petli, sumuje pow uz dla wydzielen - wymagane do
         # wpisania wartosci do tabeli F_SUBAREA
